requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
bcrypt==4.0.1
boto3==1.35.25
botocore==1.35.25
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
cssselect2==0.7.0
fastapi==0.104.1
fonttools==4.54.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
loguru==0.7.2
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
pillow==10.4.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydyf==0.10.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
RapidFuzz==3.13.0
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
tinycss2==1.3.0
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.24.0.post1
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3
pandas==2.2.3
numpy==2.2.2
